Selector contains years with events
